Adopt version 2.8.5 of gson
