+++ pom.xml
@@ -26,7 +26,7 @@
         <jax-rs.version>2.0.1</jax-rs.version>
         <log4j.version>1.2.17</log4j.version>
         <slf4j.version>1.7.2</slf4j.version>
-        <gson.version>2.3.1</gson.version>
+        <gson.version>2.8.5</gson.version>
         <snakeyaml.version>1.9</snakeyaml.version>
         <semver4j.version>2.1.0</semver4j.version>
     </properties>
